[move] 잘못된 경로에 있던 main package ui아래로 이동
