+++ app/src/main/java/com/ssu/better/presentation/ui/login/LoginScreen.kt
@@ -51,7 +51,7 @@ fun LoginScreen(
                     navController.navigate(Screen.OnBoard.route)
                 }
                 else -> {
-                    navController.navigate(Screen.Main.route) {
+                    navController.navigate(Screen.Home.route) {
                         popUpTo(Screen.Login.route) {
                             inclusive = true
                         }
