+++ app/src/main/java/com/ssu/better/presentation/ui/main/home/SampleScreen.kt
@@ -1,10 +1,10 @@
-package com.ssu.better.ui.theme.main.search
+package com.ssu.better.presentation.ui.main.home
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 
 @Composable
-fun SearchScreen(navHostController: NavHostController) {
-    Text(text = "search")
+fun SampleScreen(navHostController: NavHostController) {
+    Text(text = "sample")
 }
