+++ app/src/main/java/com/ssu/better/presentation/ui/main/mypage/MyPageScreen.kt
@@ -1,10 +1,10 @@
-package com.ssu.better.ui.theme.main.home
+package com.ssu.better.presentation.ui.main.mypage
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 
 @Composable
-fun SampleScreen(navHostController: NavHostController) {
-    Text(text = "sample")
+fun MyPageScreen(navHostController: NavHostController) {
+    Text(text = "mypage")
 }
